+++ src/Translatable/TranslatableServiceProvider.php
@@ -1,4 +1,6 @@
-<?php namespace Dimsav\Translatable;
+<?php
+
+namespace Dimsav\Translatable;
 
 use Illuminate\Support\ServiceProvider;