+++ tests/models/CityTranslation.php
@@ -1,8 +1,11 @@
-<?php namespace Dimsav\Translatable\Test\Model;
+<?php
+
+namespace Dimsav\Translatable\Test\Model;
 
 use Illuminate\Database\Eloquent\Model as Eloquent;
 
 class CityTranslation extends Eloquent
 {
     public $timestamps = false;
+    public $fillable = ['name'];
 }