+++ tests/models/CountryStrict.php
@@ -1,4 +1,6 @@
-<?php namespace Dimsav\Translatable\Test\Model;
+<?php
+
+namespace Dimsav\Translatable\Test\Model;
 
 use Dimsav\Translatable\Translatable;
 use Illuminate\Database\Eloquent\Model as Eloquent;
@@ -20,7 +22,7 @@ class CountryStrict extends Eloquent
      *
      * @var string Class containing the translation
      */
-    public $translationModel = 'Dimsav\Translatable\Test\Model\CountryTranslation';
+    public $translationModel = 'Dimsav\Translatable\Test\Model\StrictTranslation';
 
     /**
      * @var string Foreign key for the translation relationship