+++ tests/models/CountryTranslation.php
@@ -1,8 +1,12 @@
-<?php namespace Dimsav\Translatable\Test\Model;
+<?php
+
+namespace Dimsav\Translatable\Test\Model;
 
 use Illuminate\Database\Eloquent\Model as Eloquent;
 
 class CountryTranslation extends Eloquent
 {
     public $timestamps = false;
+
+    protected $fillable = ['name'];
 }