+++ tests/models/Food.php
@@ -1,4 +1,6 @@
-<?php namespace Dimsav\Translatable\Test\Model;
+<?php
+
+namespace Dimsav\Translatable\Test\Model;
 
 use Dimsav\Translatable\Translatable;
 use Illuminate\Database\Eloquent\Model as Eloquent;
@@ -32,18 +34,11 @@ class Food extends Eloquent
     public $translationForeignKey;
 
     /**
-     * Add your translated attributes here if you want
-     * fill them with mass assignment.
-     *
-     * @var array
-     */
-    public $fillable = ['name'];
-
-    /**
      * The database field being used to define the locale parameter in the translation model.
      * Defaults to 'locale'.
      *
      * @var string
      */
     public $localeKey;
+    public $fillable = ['id'];
 }