+++ tests/models/PersonTranslation.php
@@ -4,7 +4,7 @@ namespace Dimsav\Translatable\Test\Model;
 
 use Illuminate\Database\Eloquent\Model as Eloquent;
 
-class FoodTranslation extends Eloquent
+class PersonTranslation extends Eloquent
 {
     public $timestamps = false;
     public $fillable = ['name'];