+++ tests/seeds/AddFreshSeeds.php
@@ -1,17 +1,13 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Dimsav\Translatable\Test\Model\Country;
-use Dimsav\Translatable\Test\Model\CountryTranslation;
 use Dimsav\Translatable\Test\Model\City;
+use Dimsav\Translatable\Test\Model\Country;
 use Dimsav\Translatable\Test\Model\CityTranslation;
+use Dimsav\Translatable\Test\Model\CountryTranslation;
 
-class AddSeeds extends Migration
+class AddFreshSeeds
 {
-    /**
-     * Run the migrations.
-     */
-    public function up()
+    public function run()
     {
         $countries = [
             ['id' => 1, 'code' => 'gr'],
@@ -88,12 +84,5 @@ class AddSeeds extends Migration
             $translation->name = $data['name'];
             $translation->save();
         }
-    }
-
-    /**
-     * Reverse the migrations.
-     */
-    public function down()
-    {
     }
 }